Fix methods that get last letters of words
Because the methods getting last letters required the index of the last
letter to be greater than the index of the first letter, an error
occurred when an added cell block resulted in a one-letter word.

These methods now allow the last letter's index to be equal to the first
letter's index. In other words, one-letter words are permitted.

It is rare that the NYT includes a word less than three letters in one
of its puzzles, so I intend to implement a feature that offers the user
the opportunity to check that the puzzle follows the NYT's crossword
puzzle guidelines