Add word lists for 'Fill' DashboardPage
Because there are so many words, in order to cut down on time, I split
the word list into separate files based on word length and pre-formatted
them as json objects.

Word scores came with the word list and are subjectively assigned by
contributors of the word list. The lower the score is, the less willing
one should be to include that word in their puzzle

Word list source:
https://raw.githubusercontent.com/Crossword-Nexus/
collaborative-word-list/main/xwordlist.dict