Add method to minimize size of argsArr
Previously, an args obj was pushed to argsArr every time wordMatchIndex changed. Now, when trying the next wordMatchIndex, this function slices argsArr, excluding the most recent args obj, and this sliced argsArr is passed as an argument to the autofillGrid function. When an args obj is then pushed to argsArr at the beginning of the autofillGrid function, it is as though only the args obj's wordMatchIndex value updated. The size of the array does not increase. argsArr only holds one args obj for each wordToFill. This has made the debugging process easier, as there is less data to look through, and this seems to have also decreased the time it takes to fill the grid