Refactor autofillGrid to be async and more concise
Also, to prevent memory errors, in addition to backtrack being called if wordToFill has zero wordMatches, it will now also be called if wordMatchIndex > 100. This seems to have helped reduce the occurrence of errors and decrease the time it takes to fill the grid. The number was decided on after some trial and error, so that may still need to be adjusted at some point. Or a different means of preventing testing too many wordMatches may prove to be a better solution to the problem